+++ view/frontend/web/js/customer.js
@@ -4,13 +4,13 @@ define([
     'domReady!'
 ], function (_, customerData) {
     'use strict';
-    var _learnq = window._learnq || [];
+    var klaviyo = window.klaviyo || [];
 
     customerData.getInitCustomerData().done(function () {
         var customer = customerData.get('customer')();
 
-        if(_.has(customer, 'email') && customer.email && !_learnq.isIdentified()) {
-            _learnq.identify({
+        if(_.has(customer, 'email') && customer.email && !klaviyo.isIdentified()) {
+            klaviyo.identify({
                 '$email': customer.email,
                 '$first_name': _.has(customer, 'firstname') ? customer.firstname : '',
                 '$last_name':  _.has(customer, 'lastname') ? customer.lastname : ''