+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+#### Changed
+- Updated the legacy `_learnq` js object to the new `klaviyo` js object.
+
 ### [4.0.9] - 2023-01-03
 ### Changed
 - Updated default SMS consent language